Continued development.
Location and organization are now determined strictly from user profile
and input configuration. User is no longer prompted.

Compute resource is determined based on location and organization. If a
single is returned that is selected, else user is prompted. Would like
to determine a way to programatically determine the compute resource.